pyproject.toml
[tool.poetry]
name = "ontario-driving-school-manager"
version = "0.1.0"
description = "A comprehensive management system for Ontario driving schools"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "ontario_driving_school_manager", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
PyQt5 = "^5.15.9"
SQLAlchemy = "^2.0.23"
pydantic = "^2.5.2"
alembic = "^1.12.1"
python-dotenv = "^1.0.0"
PyYAML = "^6.0.1"
loguru = "^0.7.2"
pydantic-settings = "^2.9.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 